app_v4/requirements.txt

numpy
librosa
soundfile
pyloudnorm; python_version >= "3.8"
python-dotenv
requests
shazamio; python_version >= "3.8"
